+++ src/Config/Services.php
@@ -9,13 +9,34 @@ use CodeIgniter\HTTP\UserAgent;
 use Config\App;
 use Config\Services as AppServices;
 use Config\Toolbar as ToolbarConfig;
+use Config\View as ViewConfig;
 use Michalsn\CodeIgniterHtmx\Debug\Toolbar;
 use Michalsn\CodeIgniterHtmx\HTTP\IncomingRequest;
 use Michalsn\CodeIgniterHtmx\HTTP\RedirectResponse;
 use Michalsn\CodeIgniterHtmx\HTTP\Response;
+use Michalsn\CodeIgniterHtmx\View\View;
 
 class Services extends BaseService
 {
+    /**
+     * The Renderer class is the class that actually displays a file to the user.
+     * The default View class within CodeIgniter is intentionally simple, but this
+     * service could easily be replaced by a template engine if the user needed to.
+     *
+     * @return View
+     */
+    public static function renderer(?string $viewPath = null, ?ViewConfig $config = null, bool $getShared = true)
+    {
+        if ($getShared) {
+            return static::getSharedInstance('renderer', $viewPath, $config);
+        }
+
+        $viewPath = $viewPath ?: config('Paths')->viewDirectory;
+        $config ??= config('View');
+
+        return new View($config, $viewPath, AppServices::locator(), CI_DEBUG, AppServices::logger());
+    }
+
     /**
      * Returns the current Request object.
      *