+++ tests/_support/Views/layout.php
@@ -1,0 +1,3 @@
+Page top
+<?= $this->renderSection('content') ?>
+<?php $this->fragment('sample0'); ?>Page bottom<?php $this->endFragment(); ?>