+++ tests/_support/Views/view_fragment.php
@@ -1,0 +1,8 @@
+Fragment 1 header
+<?php $this->fragment('sample1'); ?>
+<?= $testString1 ?? 'Hello' ?>, fragment1!
+<?php $this->endFragment(); ?>
+Fragment 2 header
+<?php $this->fragment('sample2'); ?>
+<?= $testString2 ?? 'Hello' ?>, fragment2!
+<?php $this->endFragment(); ?>