+++ tests/_support/Views/view_fragment_error.php
@@ -1,0 +1,2 @@
+<?= $testString1 ?? 'Hello' ?>, fragment1!
+<?php $this->endFragment(); ?>