+++ tests/_support/Views/view_fragment_orig.php
@@ -1,0 +1,4 @@
+Fragment 1 header
+<?= $testString1 ?? 'Hello' ?>, fragment1!
+Fragment 2 header
+<?= $testString2 ?? 'Hello' ?>, fragment2!