+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - When an HTTP error response occurs, display it in the modal.
+- Add `view_fragment()` function to return only fragments of the view.
 
 ## [1.0.0-beta.2](https://github.com/michalsn/codeigniter-htmx/compare/v1.0.0-beta...v1.0.0-beta.2) - 2022-11-22